+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# 1.1.0
+- Added `Error::output` which returns an `Option<String>`, the output produced before the error occurred (if applicable)
+- Try to avoid an allocation when creating output
+
 # 1.0.0
 - Crate release